+++ app/controllers/blog_controller.rb
@@ -3,18 +3,14 @@
 # Publicly available and published blog posts
 class BlogController < ApplicationController
   before_action :find_broadcast_or_redirect, only: [:show]
+  before_action :set_author, only: [:blog]
+  before_action :set_category, only: [:blog]
 
   def blog
     broadcast_scope = Broadcast.current.published.order(publish_date: :desc, id: :desc)
-    unless params[:author].blank?
-      @author = User.current.with_name(params[:author]).first
-      broadcast_scope = broadcast_scope.where(user: @author)
-    end
-    unless params[:category].blank?
-      broadcast_scope = broadcast_scope.joins(:category).merge(Category.current.where(slug: params[:category]))
-    end
-    @broadcasts = broadcast_scope.page(params[:page]).per(40)
-
+    broadcast_scope = broadcast_scope.where(user: @author) if @author
+    broadcast_scope = broadcast_scope.where(category: @category) if @category
+    @broadcasts = broadcast_scope.page(params[:page]).per(10)
     respond_to do |format|
       format.html
       format.atom
@@ -24,7 +20,12 @@ class BlogController < ApplicationController
   def show
     @author = @broadcast.user
     @page = (params[:page].to_i > 1 ? params[:page].to_i : 1)
-    @replies = @broadcast.replies.includes(:broadcast).where(reply_id: nil).page(@page).per(Reply::REPLIES_PER_PAGE)
+
+    @order = scrub_order(Reply, params[:order], 'points desc')
+    if ['points', 'points desc'].include?(params[:order])
+      @order = params[:order]
+    end
+    @replies = @broadcast.replies.points.includes(:broadcast).where(reply_id: nil).reorder(@order).page(params[:page]).per(Reply::REPLIES_PER_PAGE)
     @broadcast.increment! :view_count
   end
 
@@ -33,5 +34,15 @@ class BlogController < ApplicationController
   def find_broadcast_or_redirect
     @broadcast = Broadcast.current.published.find_by_slug params[:slug]
     redirect_to blog_path unless @broadcast
+  end
+
+  def set_author
+    return if params[:author].blank?
+    @author = User.current.where('lower(username) = ?', params[:author].downcase).first
+  end
+
+  def set_category
+    return if params[:category].blank?
+    @category = Category.current.where('lower(slug) = ?', params[:category].downcase).first
   end
 end