+++ db/migrate/20160621215417_add_index_to_dataset_file_audits.rb
@@ -1,0 +1,5 @@
+class AddIndexToDatasetFileAudits < ActiveRecord::Migration
+  def change
+    add_index :dataset_file_audits, :remote_ip
+  end
+end